requirements.txt
amqp==5.2.0
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.0
bs4==0.0.2
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
drf-yasg==1.21.9
eventlet==0.37.0
flower==2.0.1
greenlet==3.1.0
gunicorn==23.0.0
humanize==4.11.0
idna==3.10
inflection==0.5.1
kombu==5.4.0
latex2mathml==3.77.0
lxml==5.3.0
numpy==1.26.4
packaging==24.1
pandas==2.2.3
phonenumbers==8.13.47
pillow==10.4.0
prometheus_client==0.21.1
prompt_toolkit==3.0.47
psycopg2==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
